tracing: Fix hist trigger totals display
Using %lu instead of %llu to display a u64 causes really strange
results e.g. entries and drop values swapped in the output.

+++ kernel/trace/trace_events_hist.c
@@ -1011,10 +1011,9 @@ static void hist_trigger_show(struct seq_file *m,
 		ret = n_entries;
 		n_entries = 0;
 	}
-
-	seq_printf(m, "\nTotals:\n    Hits: %lu\n    Entries: %u\n    Dropped: %lu\n",
-		   atomic64_read(&hist_data->total_hits),
-		   n_entries, atomic64_read(&hist_data->drops));
+	seq_printf(m, "\nTotals:\n    Hits: %llu\n    Entries: %u\n    Dropped: %llu\n",
+		   (u64)atomic64_read(&hist_data->total_hits),
+		   n_entries, (u64)atomic64_read(&hist_data->drops));
 }
 
 static int hist_show(struct seq_file *m, void *v)
